fix: expose types for showMessage handlers
